gpu: pvr: Use DVFS locks for min bus freq
The GPU will behave badly if the bus freq is set too low.

This will happen when the CPU downclocks to 100 MHz.

Previously this was done by changing the cpufreq
min policy to 200 MHz, but this could sometimes
"jam" the min policy to 200 MHz.  This PM code
is already architecture-dependent, let's use
the DVFS locks the architecture cpufreq code
offers.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>